pyproject.toml
[tool.poetry]
name = "django-avaliacao-01"
version = "0.1.0"
description = ""
authors = ["igorsenna <engenheirostrauss2@gmail.com>"]
readme = "README.md"
packages = [{include = "django_avaliacao_01"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
pytest-django = "^4.5.2"
django-factory-boy = "^1.0.0"
coverage = "^7.2.3"
psycopg2-binary = "^2.9.6"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
bootstrap4 = "^0.1.0"
pytz = "^2023.3"
model-bakery = "^1.11.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.1"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "carros_cidade.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]